Subo el desafío 15 (la última práctica integradora), en su estado final (como punto de partida para empezar a trabajar sobre la entrega final del proyecto)

+++ src/controllers/products.controller.js
@@ -1,8 +1,9 @@
 import ProductService from '../services/products.service.js'
+import config from "../config.js"
 
 let productService = new ProductService
 
-const getProducts = async (req, res) => {
+const getProducts = async (req, res, next) => {
   try {
     let limit = req.query.limit
     let page = req.query.page
@@ -28,26 +29,30 @@ const getProducts = async (req, res) => {
     res.send(response)
   }
   catch (error) {
-    res.status(400).send({status: "error", details: "There was an error"})
+    return next(error)
   }
 }
 
-const getProductById = async (req, res) => {
-  let id = req.params.pid
+const getProductById = async (req, res, next) => {
+  try {
+    let id = req.params.pid
 
-  let product = await productService.getProductById(id)
+    let product = await productService.getProductById(id)
 
-  if (!product) {
-    res.send("No se encontró el producto")
-    return
+    res.send(product)
   }
-
-  res.send(product)
+  catch(error) {
+    return next(error)
+  }
 }
 
-const addProduct = async (req, res) => {
+const addProduct = async (req, res, next) => {
   try {
     let newProduct = req.body
+
+    if (req.user.email != config.ADMIN_NAME) {
+      newProduct.owner = req.user.email // Se liga el mail del usuario como owner del producto 
+    }
 
     await productService.addProduct(newProduct)
     
@@ -57,28 +62,55 @@ const addProduct = async (req, res) => {
     res.send({status: "success"})
   }
   catch(error) {
-    res.status(400).send({status: "failure", details: error.message})
+    return next(error)
   }
 }
 
-const updateProduct = async (req, res) => {
-  let id = req.params.pid
-  let newProduct = req.body
+const updateProduct = async (req, res, next) => {
+  try {
+    let id = req.params.pid
 
-  await productService.updateProduct(id, newProduct)
+    let product = await productService.getProductById(id) // Si el producto no existe, tira error
+    
+    // El usuario debe ser "admin" o el "owner" del producto
+    if ( !(req.user.role === "admin" || product.owner === req.user.email) ) {
+      return res.status(403).
+      send({ status: "failure", details: "You don't have access. You are not the product owner" })
+    }
 
-  res.send({status: "success"})
+    let newProduct = req.body
+
+    await productService.updateProduct(id, newProduct)
+
+    res.send({status: "success"})
+  }
+  catch (error) {
+    return next(error)
+  }
 }
 
-const deleteProduct = async (req, res) => {
-  let id = req.params.pid
-  
-  await productService.deleteProduct(id)
+const deleteProduct = async (req, res, next) => {
+  try {
+    let id = req.params.pid
 
-  const products = await productService.getProducts()
-  req.socketServer.sockets.emit('update-products', products) // Para que se actualicen los productos en tiempo real
+    let product = await productService.getProductById(id) // Si el producto no existe, tira error
+    
+    // El usuario debe ser "admin" o el "owner" del producto
+    if ( !(req.user.role === "admin" || product.owner === req.user.email) ) {
+      return res.status(403).
+      send({ status: "failure", details: "You don't have access. You are not the product owner" })
+    }
 
-  res.send({status: "success"})
+    await productService.deleteProduct(id)
+
+    const products = await productService.getProducts()
+    req.socketServer.sockets.emit('update-products', products) // Para que se actualicen los productos en tiempo real
+
+    res.send({status: "success"})
+  }
+  catch(error) {
+    return next(error)
+  }
 }
 
 export default {
